pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "kalitron-telegram-bot"
version = "0.1.0"
description = "Telegram bot MVP integrated with ai-gateway."
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27,<0.29",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "python-telegram-bot>=22,<23",
]

[project.scripts]
kalitron-telegram-bot = "kalitron_telegram_bot.main:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.24,<1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
